+++ CHANGELOG.md
@@ -49,8 +49,9 @@
 ### 1.1.1
 
 - Updated to NeoForge 21.0.53-beta
-- Added 4 new Genes: Infested, Oozing, Weaving, and Wind Charged
 - Added EMI support
+- Added 4 new Genes: Infested, Oozing, Weaving, and Wind Charged
+- Fixed a recipe conflict between Panacea and Zombify Villager potions, Panacea now requires a Regeneration Helix rather than Emerald Heart
 - Sorted Genes in several places (Sorted firstly by regular/mutation/negative, then by id)
 - Fixed a typo in a tooltip
 - Forgot the missing : in several places in the mod name